Rename image to satchless_image and fix tests
* Upgrade the test runner to be compatible with django 1.6
* Delete the satchless.contrib.payment.django_payments_provider module
  because it does not work (has unresolvable imports which don't
  correspond to current functionality, making it totally unusable)
