+++ doc/intro/installation.rst
@@ -31,7 +31,7 @@ Include neccessary applications in your ``INSTALLED_APPS`` tuple in Django
         'django.contrib.admin',
         'mptt',
         'satchless.product',
-        'satchless.image',
+        'satchless.satchless_image',
 
         # You may skip the following ones when running just a product catalog:
         'satchless.contact',
@@ -54,7 +54,7 @@ just examples and you may change them freely.::
     urlpatterns = patterns('',
         url(r'^products/', include(product_app.urls)),
         url(r'^contact/', include('satchless.contact.urls')),
-        url(r'^image/', include('satchless.image.urls')),
+        url(r'^image/', include('satchless.satchless_image.urls')),
         url(r'^cart/', include(cart_app.urls)),
         url(r'^checkout/', include(order_app.urls)),
     )
