+++ examples/demo/categories/models.py
@@ -2,7 +2,7 @@
 from django.conf import settings
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
-from satchless.image.models import Image
+from satchless.satchless_image.models import Image
 
 from localeurl.models import reverse
 from mothertongue.models import MothertongueModelTranslate
