+++ examples/demo/urls.py
@@ -15,7 +15,7 @@ urlpatterns = patterns('',
     url(r'^payment/failed/(?P<order_token>\w+)/$', core.views.payment_failed, name='payment-failed'),
     url(r'^products/', include(product_app.urls)),
     url(r'^contact/', include('satchless.contact.urls')),
-    url(r'^image/', include('satchless.image.urls')),
+    url(r'^image/', include('satchless.satchless_image.urls')),
     url(r'^cart/', include(cart_app.urls)),
     url(r'^wishlist/', include(wishlist_app.urls)),
     url(r'^order/', include(order_app.urls)),
