+++ examples/dummy/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-from satchless.image.models import Image
+from satchless.satchless_image.models import Image
 from satchless.product.models import ProductAbstract, Variant
 
 class Dummy(ProductAbstract):
