+++ satchless/__init__.py
@@ -4,4 +4,4 @@ Satchless
 The Django e-commerce framework for perfectionists who also care about code
 quality.
 '''
-VERSION = (0, 0, 19)
+VERSION = (0, 0, 20)
