--- satchless/contrib/payment/django_payments_provider/views.py
@@ -1,1 +1,0 @@
-# Create your views here.
