+++ satchless/satchless_image/templatetags/satchless_image.py
